% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{adapt_tsne}
\alias{adapt_tsne}
\title{wrapper for rtsne function that will set perplexity to 1 if an error occurs}
\usage{
adapt_tsne(dist)
}
\arguments{
\item{dist}{input distance matrix, must be type dist}
}
\value{
rtsne result
}
\description{
wrapper for rtsne function that will set perplexity to 1 if an error occurs
}
\examples{
#soon
}
