% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ccrepe_to_geomnet}
\alias{ccrepe_to_geomnet}
\title{filter ccrepe object and prepare for geomnet plotting}
\usage{
ccrepe_to_geomnet(ccrepe.obj, pcut = 0.05, spearcut = 0.6)
}
\arguments{
\item{ccrepe.obj}{an object returned by the function ccrepe()}

\item{pcut}{The pvalue cutoff, all correlations with pvalues above this will be excluded}

\item{spearcutThe}{spearman correlation coefficient cutoff.  All correlations with spear values less than this will be removed
Needs improvement as this function cannot only return strong negative correlations...}
}
\value{
Returns an edgelist dataframe, should be basically ready for use with geomnet
}
\description{
filter ccrepe object and prepare for geomnet plotting
}
\examples{
coming soon
}
