% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{cum_dist_to_targets}
\alias{cum_dist_to_targets}
\title{Cumulative distance to targets in a distance matrix}
\usage{
cum_dist_to_targets(tmat, pattern)
}
\arguments{
\item{tmat}{MATRIX representation of distance matrix}

\item{pattern}{that defines your targets, passed to grep}
}
\value{
returns a named vector of cumulative distance to your specified targets (for every entry)
}
\description{
Cumulative distance to targets in a distance matrix
}
\examples{
#Soon
}
