% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{darken}
\alias{darken}
\title{Return a darker hue of a color}
\usage{
darken(color, factor = 1.4)
}
\arguments{
\item{factor}{}
}
\description{
I did not write this function, taken from:
https://gist.github.com/Jfortin1/72ef064469d1703c6b30
}
