% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Deseq.quickplot}
\alias{Deseq.quickplot}
\title{Quickly plot Deseq2 results}
\usage{
Deseq.quickplot(DeSeq.object, phyloseq.object, pvalue = 0.05, contrast.vector,
  taxlabel = "Genus")
}
\arguments{
\item{DeSeq.object}{Deseq object that the desired pairwise contrast exists in}

\item{phyloseq.object}{phyloseq object containing the taxonomy table associated with the OTUs in the Deseq object}

\item{pvalue}{pvalue to filter results table with}

\item{contrast.vector}{a vector of length 3, 1st entry is the name of the factor containing the groups, second 2 are group names}

\item{taxlabel}{the taxonomy label to use for the plot labels}
}
\value{
returns a list containing: [[1]] a ggplot object, [[2]] a dataframe containing the significantly differentially abundant features
}
\description{
Quickly plot Deseq2 results
}
\examples{
coming soon
}
