% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DESeq_cov_asso}
\alias{DESeq_cov_asso}
\title{DESeq2 covariate associations with OTUs}
\usage{
DESeq_cov_asso(
  phyloseq_obj,
  day,
  tissue,
  covariate,
  shrink_type = "apeglm",
  cookscut = TRUE,
  treatment = FALSE,
  scale_cov = TRUE,
  p.plot = 0.05,
  plot_lab = "Genus",
  l2fc_plot = 0.25
)
}
\arguments{
\item{phyloseq_obj}{A phyloseq object, must have tax table}

\item{day}{Day to prune the phyloseq object to}

\item{tissue}{tissue to prune the phyloseq object to}

\item{covariate}{Covariate in sample data to calculate associations for, as a string.}

\item{shrink_type}{DESeq2 shrink type}

\item{cookscut}{Apply cooks cutoff?}

\item{treatment}{include 'treatment' in formula?}

\item{scale_cov}{scale the covariate?}

\item{p.plot}{pvalue to filter plot by}

\item{plot_lab}{taxonomic level to use for plotting}

\item{l2fc_plot}{lfc2 to filter plot by}
}
\value{
returns a list of lenth 2, 1st is plot 2nd is data
}
\description{
DESeq2 covariate associations with OTUs
}
\examples{
#soon
}
