% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DESeq_difabund}
\alias{DESeq_difabund}
\title{DESeq2 differential abundance for all treatments compared to 'Control'}
\usage{
DESeq_difabund(
  phyloseq,
  day,
  tissue,
  scientific = TRUE,
  shrink_type = "normal",
  alpha = 0.1,
  cooks_cut = FALSE,
  pAdjustMethod = "BH"
)
}
\arguments{
\item{phyloseq}{phyloseq object}

\item{day}{day of interest (used to subset phyloseq object)}

\item{tissue}{tissue of interest (used to subset phyloseq object)}

\item{scientific}{change pvals to sci notation?}

\item{shrink_type}{what type of lfc shrinkage? passed to DESeq2's lfc_shrink.
should be one of 'normal', 'ashr', 'apeglm'.}

\item{alpha}{passed to DESeq2 results() function.}

\item{cooks_cut}{apply cooks cutoff?}

\item{pAdjustMethod}{what type of p.adjustment to apply?}
}
\value{
returns a dataframe containing all the significantly differentially abundant taxa for
each treatment in your experiment compared to the 'Controls'
}
\description{
DESeq2 differential abundance for all treatments compared to 'Control'
}
\examples{
#soon
}
