% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_mothur_tax}
\alias{extract_mothur_tax}
\title{Extract mothur SILVA formatted taxonomy}
\usage{
extract_mothur_tax(filename)
}
\arguments{
\item{filename}{A taxonomy file output directly from mothur}
}
\value{
a dataframe with the OTU taxonomy classifications split into all the various taxonomic levels
}
\description{
Extract mothur SILVA formatted taxonomy
}
\examples{
#soon
}
