% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gather_nodes}
\alias{gather_nodes}
\title{Collect type information for nodes in a network}
\usage{
gather_nodes(x, typ = NA)
}
\arguments{
\item{x}{a dataframe with columns as measured variables that could be nodes in a correlation network}

\item{typ}{either a single text value or a text vector of length equal to the number of columns
describing the type(s) of these potential nodes}
}
\value{
returns a two column dataframe. The column 'node' is populated with the column names from the original dataframe.
 The 'type' column is user supplied.
}
\description{
Collect type information for nodes in a network
}
\examples{
coming soon
}
