% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{iter_plots}
\alias{iter_plots}
\title{plots the output of iterative_NMDS}
\usage{
iter_plots(iter_res, iter, pattern, labels = TRUE, mysize = 1)
}
\arguments{
\item{iter_res}{the full list output by iterative_NMDS}

\item{iter}{the iteration you want to plot, should be an integer}

\item{pattern}{the pattern that will identify objects you are interested in, passed to grep}

\item{labels}{should we label things in the plots?}

\item{mysize}{what size do you want the points you are interested in to have?}
}
\description{
plots the output of iterative_NMDS
}
\examples{
#soon
}
