% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{iterative_NMDS}
\alias{iterative_NMDS}
\title{Iterative removal of objects in a distance matrix}
\usage{
iterative_NMDS(in_dist, trymax = 50, iterations = 20, maxit = 1000,
  autotransform = FALSE, exclusion_prob = 0.95, parallel = 8,
  targ_pat = "SX", rem_type = "total", nmds = FALSE)
}
\arguments{
\item{in_dist}{input distance matrix (should be type dist)}

\item{trymax}{passed to metaMDS}

\item{iterations}{number of removal iterations}

\item{maxit}{passed to metaMDS}

\item{autotransform}{passed to metaMDS}

\item{exclusion_prob}{passed to quantile(), cumulative distances falling above this cutoff will be removed
each iteration, default = 0.95}

\item{parallel}{passed to metaMDS}

\item{targ_pat}{passed to grep, used to identify targets (things you want to keep)}

\item{rem_type}{determines how to remove objects from distance matrix, 'total' removes points
with the highest cumulative distance from all other objects.  "cum_targ" removes objects with the
highest cumulative distance from your specified targets. 'min' will remove points with the highest
cumulative minimum distance from your targets}

\item{nmds}{Should nmds be run? (can be slow)}
}
\value{
returns a list of lists, one for each iteration, the list for each iteration
 contains: [[1]] a dataframe for plotting, [[2]] nmds plot if specified,
[[3]] a distance matrix with objects removed, [[4]] a tsne plot
}
\description{
Iterative removal of objects in a distance matrix
}
\examples{
#soon
}
