% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lighten}
\alias{lighten}
\title{Return a lighter hue of a color}
\usage{
lighten(color, factor = 1.4)
}
\arguments{
\item{color}{color to lighten}

\item{factor}{how much lightening to do}
}
\value{

}
\description{
I did not write this function, taken from:
https://gist.github.com/Jfortin1/72ef064469d1703c6b30
}
\examples{
#soon
}
