% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{min_dist_to_targets}
\alias{min_dist_to_targets}
\title{Minimum distance to targets for each other entry in a distance matrix}
\usage{
min_dist_to_targets(tmat, pattern)
}
\arguments{
\item{tmat}{MATRIX representation of distance matrix}

\item{pattern}{that defines your targets, passed to grep}
}
\value{
returns a named vector of minimum distance to your specified targets (for every entry)
}
\description{
Minimum distance to targets for each other entry in a distance matrix
}
\examples{
#soon
}
