% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{NMDS_ellipse}
\alias{NMDS_ellipse}
\title{Generate NMDS points, group centroids and standard error ellipses}
\usage{
NMDS_ellipse(metadata, OTU_table, grouping_set, distance_method = "bray",
  rand_seed = 77777, MDS_trymax = 1000, autotransform = FALSE,
  wascores = TRUE, expand = FALSE)
}
\arguments{
\item{metadata}{A dataframe containing the metadata relating to the samples in your OTU table}

\item{OTU_table}{an OTU table with samples as rows and taxa as columns}

\item{grouping_set}{a column name from your metadata.  This column should contain information
about how your samples should be grouped for centroid and ellipse calculation}

\item{distance_method}{a single text value that describes the desired distance metric to use, must be
accepted by the function vegdist()}

\item{rand_seed}{numeric value to use as a random seed (for reproducability?)}

\item{MDS_trymax}{number of iterations to use for the metaMDS call, default is 1000}

\item{autotransform}{passed to vegan's metaMDS, should community data be transformed?
Should be TRUE/FALSE.  Default = FALSE}

\item{wascores}{passed to vegan's metaMDS, should species scores be calculated?
Should be TRUE/FALSE.  Default = TRUE}

\item{expand}{passed to vegan's metaMDS, should species scores be expanded?}
}
\value{
Returns a list containing 3 items. [[1]] = input metadata with NMDS coordinates, group centroid coordinates for each sample.
  [[2]] = a dataframe containing information about the standard error confidence ellipses for each group.  This datafame
  contains 100 points per group and is suitable for using with geom_path().
  [[3]] = the ordination object created by the vegan function metaMDS().
}
\description{
Generate NMDS points, group centroids and standard error ellipses
}
\examples{
coming soon
}
