% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{otu_tax_labels}
\alias{otu_tax_labels}
\title{swap OTU designations with 'genus:OTU'}
\usage{
otu_tax_labels(mothur_tax)
}
\arguments{
\item{mothur_tax}{A mothur taxonomy file read in by "extract_mothur_tax()"}
}
\value{
Returns a named vector.  The values of this vector are new names for each OTU.  These are made by concatenating the entry in the
 'Genus' column with the entry in the 'OTU' column separated by a ':'.  This named vector can then be used to get the new names for
 any subset of these OTUs. I use this function when I've got some plot involving a few OTUs and I want to label them with both an OTU number and
 some taxonomy information.
}
\description{
swap OTU designations with 'genus:OTU'
}
