% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pairwise.adonis}
\alias{pairwise.adonis}
\title{this function taken from https://www.researchgate.net/post/How_can_I_do_PerMANOVA_pairwise_contrasts_in_R}
\usage{
pairwise.adonis(
  x,
  factors,
  sim.method = "bray",
  p.adjust.m = "none",
  permutations = 999
)
}
\arguments{
\item{x}{An OTU table with taxa as columns and samples as rows}

\item{factors}{a vector containing the groups you want compared}

\item{sim.method}{distance metric to use for (dis)similarity values, default is 'bray'}

\item{p.adjust.m}{method to use for pvalue correction.}

\item{permutations}{number of permutations to use}
}
\value{
returns a dataframe containing the PERMANOVA rsq, fstat, and pvalue for each pariwise comparison
}
\description{
this function taken from https://www.researchgate.net/post/How_can_I_do_PerMANOVA_pairwise_contrasts_in_R
}
\examples{
#soon
}
