% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prune_graph}
\alias{prune_graph}
\title{Remove small unconnected subnetworks from a network}
\usage{
prune_graph(fortified.edgelist, node.dataframe, min.vert = 1)
}
\arguments{
\item{fortified.edgelist}{This dataframe should be a network stored as a "fortified.edgelist" (see geomnet package for details?)}

\item{node.dataframe}{This dataframe should be one returned by gather_nodes() or a bunch of these concatenated together}

\item{min.vert}{A number specifiing the minimum number of verticies that can be in any subnetwork within the larger network}
}
\value{
Returns a network fit for plotting with geomnet in which small subnetworks with less than the specified number have been removed
}
\description{
Remove small unconnected subnetworks from a network
}
\examples{
coming soon
}
