% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qpcr_DDCT}
\alias{qpcr_DDCT}
\title{CTs to log2 fold change relative to controls}
\usage{
qpcr_DDCT(data, exp_genes_indicies, housekeeping_gene_index,
  control_row_indicies, metadata_column_indicies)
}
\arguments{
\item{data}{this is the input dataframe, should have genes as columns and samples as rows, replicates should be collpsed and only one housekeeping gene
is supported.  I don't think NA's or missing data is supported...not really sure.}

\item{exp_genes_indicies}{A numeric vector containing the column indicies for the genes of experimental interest}

\item{housekeeping_gene_index}{A single numeric value defining the column index for the housekeeping gene}

\item{control_row_indicies}{A numeric vector containing the row indicies for the samples that are in the control group.  Final data will be
expressed relative to the mean of this group for each gene.}

\item{metadata_column_indicies}{Numeric column indicies defining which columns are metadata and should be returned along with the final results}
}
\value{
returns a dataframe of log2(2^DDct) values centered on the mean of the control group
}
\description{
CTs to log2 fold change relative to controls
}
