% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qpcr_DDCT_LOD}
\alias{qpcr_DDCT_LOD}
\title{CTs to log2 fold change relative to controls with LOD estimates}
\usage{
qpcr_DDCT_LOD(
  data,
  exp_genes_indicies,
  housekeeping_gene_index,
  control_row_indicies,
  metadata_column_indicies,
  LOD,
  LOD_conf = 0.999
)
}
\arguments{
\item{data}{is is the input dataframe, should have genes as columns and samples as rows, replicates should be collpsed and only one housekeeping gene
is supported.  I don't think NA's or missing data is supported...not really sure.}

\item{exp_genes_indicies}{A numeric vector containing the column indicies for the genes of experimental interest}

\item{housekeeping_gene_index}{A single numeric value defining the column index for the housekeeping gene}

\item{control_row_indicies}{A numeric vector containing the row indicies for the samples that are in the control group.  Final data will be
expressed relative to the mean of this group for each gene.}

\item{metadata_column_indicies}{Numeric column indicies defining which columns are metadata and should be returned along with the final results}

\item{LOD}{Your desired Limit of detection, all cts above this value will be replaced with NA and then this value}

\item{LOD_conf}{The confidence level for the LOD estimate}
}
\value{
returns list of length 2, 1) a dataframe of log2(2^DDct) values centered on the mean of the control group
and 2) a dataframe of LOD estimates for each gene
}
\description{
CTs to log2 fold change relative to controls with LOD estimates
}
\examples{
#soon
}
