% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rcorr_to_geomnet}
\alias{rcorr_to_geomnet}
\title{filter rcorr object and prepare for geomnet plotting}
\usage{
rcorr_to_geomnet(rcorr.list, pcut = 0.05, spearcut = 0.6)
}
\arguments{
\item{rcorr.list}{The object returned by the rcorr() function}

\item{pcut}{The pvalue cutoff, all correlations with pvalues above this will be excluded}

\item{spearcut}{The spearman correlation coefficient cutoff.  All correlations with spear values less than this will be removed
Needs improvement as this function cannot only return strong negative correlations...}
}
\value{
Returns an edgelist dataframe, should be basically ready for use with geomnet
}
\description{
filter rcorr object and prepare for geomnet plotting
}
\examples{
#soon
}
