% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{read_mash_tri}
\alias{read_mash_tri}
\title{Read in a mash output lower triangle file as a dist object}
\usage{
read_mash_tri(file)
}
\arguments{
\item{file}{A lower-triangle relaxed phylip distance matrix output by the mash triangle command}
}
\value{
Returns a distance matrix of type dist
}
\description{
Read in a mash output lower triangle file as a dist object
}
\examples{
#comming soon
}
