% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{topGO_NonModel}
\alias{topGO_NonModel}
\title{GO term enrichment from non-model organisms}
\usage{
topGO_NonModel(Int_genes, mapping_file, ont = "BP", algor = "elim",
  statistic = "Fisher")
}
\arguments{
\item{Int_genes}{Genes/proteins that are enriched in some way/group}

\item{mapping_file}{Go term mapping file containing all genes/proteins in your experiment}

\item{ont}{Which ontology should be used? CC MF or BP}

\item{algor}{which algorithm should be used in the test? see topGO package}

\item{statistic}{which test statistic should be used? see topGO package}
}
\value{
returns a table of the test results for each GO term detected in your int_genes table
}
\description{
GO term enrichment from non-model organisms
}
\examples{
#None yet.
}
