% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{topGO_wrapper}
\alias{topGO_wrapper}
\title{Wrapper for TopGO - GO term enrichment analysis}
\usage{
topGO_wrapper(
  myInterestingGenes,
  mapping_file,
  ont = "BP",
  algor = "elim",
  statistic = "Fisher",
  nodeSize = 5
)
}
\arguments{
\item{myInterestingGenes}{A vector of genes of interest}

\item{mapping_file}{a two column file, first column is the gene_ID,
second column is a comma delimited list of GO terms associated with that gene}

\item{ont}{Which GO ontology to use? options are: 'BP', 'MF', 'CC'}

\item{algor}{What algorithm to use? default is 'elim'}

\item{statistic}{what test statistic to use? default is 'Fisher'}

\item{nodeSize}{used to filter rare GO terms, default is 5.
Rare GO terms will often show up as significant as an artifact of the test methodology}
}
\value{
Returns a dataframe tests for all GO terms in your data.
You need to filter based on a reasonable pvalue
}
\description{
Wrapper for TopGO - GO term enrichment analysis
}
\examples{
#none yet
}
