% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{veganCovEllipse}
\alias{veganCovEllipse}
\title{veganCovEllipse Helper function from vegan to calculate confidence ellipses given an MDS object}
\usage{
veganCovEllipse(cov, center = c(0, 0), scale = 1, npoints = 100)
}
\arguments{
\item{cov}{.}

\item{center}{.}

\item{scale}{.}

\item{npoints}{.}
}
\value{
.
}
\description{
veganCovEllipse Helper function from vegan to calculate confidence ellipses given an MDS object
}
\examples{
...
}
