% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Deseq.quickplot}
\alias{Deseq.quickplot}
\title{Quickly plot Deseq2 results}
\usage{
Deseq.quickplot(DeSeq.object, phyloseq.object, pvalue = 0.05, contrast.vector,
  taxlabel = "Genus", colors = NULL)
}
\arguments{
\item{DeSeq.object}{Deseq object that the desired pairwise contrast exists in}

\item{phyloseq.object}{phyloseq object containing the taxonomy table associated with the OTUs in the Deseq object}

\item{pvalue}{pvalue to filter results table with}

\item{contrast.vector}{a vector of length 3, 1st entry is the name of the factor containing the groups, second 2 are group names you want to contrast}

\item{taxlabel}{the taxonomy label to use for the plot labels (needs to be present in the tax table from the phyloseq object)}

\item{colors}{optional. A vector of length 2, first value will be the color associated with the 2nd value from the contrast vector, 2nd value will be the color associated with the 3rd value from the contrast vector}
}
\value{
returns a list containing: [[1]] a ggplot object, [[2]] a dataframe containing the significantly differentially abundant features
}
\description{
Quickly plot Deseq2 results
}
\examples{
coming soon
}
