% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pairwise.adonis}
\alias{pairwise.adonis}
\title{Pairwise adonis}
\usage{
pairwise.adonis(x, factors, sim.method, p.adjust.m)
}
\arguments{
\item{x}{An OTU table with taxa as columns and samples as rows}

\item{factors}{a vector containing the groups you want compared}

\item{sim.method}{distance metric to use for (dis)similarity values, default is 'bray'}

\item{p.adjust.m}{method to use for pvalue correction.}
}
\value{
returns a dataframe containing the PERMANOVA rsq, fstat, and pvalue for each pariwise comparison
}
\description{
Pairwise adonis
}
\examples{
coming soon
}
